Refactor int test to use pytest
