+++ scripts/int_test.sh
@@ -48,13 +48,14 @@ test_service() {
   # Can't just install requirements-file, as it points to all dependencies
   pip install -r requirements-tests.txt > /dev/null
   pip install .[$service] > /dev/null 2>&1
+  pip install boto > /dev/null 2>&1
   # Restart venv - ensure these deps are loaded
   deactivate
   source ${venv_path}/bin/activate > /dev/null
   # Run tests for this service
   test_result_filename="test_results_${service}.log"
   touch $test_result_filename
-  nosetests -qxs --ignore-files="test_server\.py" --ignore-files="test_${service}_cloudformation\.py" --ignore-files="test_integration\.py" $path_to_test_file >$test_result_filename 2>&1
+  pytest -sv --ignore-glob="**/test_server.py" --ignore-glob="**/test_*_cloudformation.py" --ignore-glob="**/test_*_integration.py" $path_to_test_file >$test_result_filename 2>&1
   RESULT=$?
   if [[ $RESULT != 0 ]]; then
     echo -e "Tests for ${service} have failed!\n"
