+++ tests/test_ec2/test_iam_integration.py
@@ -307,40 +307,3 @@ def test_invalid_disassociate():
         client.disassociate_iam_instance_profile(AssociationId="fake",)
     ex.value.response["Error"]["Code"].should.equal("InvalidAssociationID.NotFound")
     ex.value.response["Error"]["Message"].should.contain("An invalid association-id of")
-
-
-@mock_ec2
-@mock_cloudformation
-def test_cloudformation():
-    dummy_template_json = {
-        "AWSTemplateFormatVersion": "2010-09-09",
-        "Resources": {
-            "InstanceProfile": {
-                "Type": "AWS::IAM::InstanceProfile",
-                "Properties": {"Path": "/", "Roles": []},
-            },
-            "Ec2Instance": {
-                "Type": "AWS::EC2::Instance",
-                "Properties": {
-                    "IamInstanceProfile": {"Ref": "InstanceProfile"},
-                    "KeyName": "mykey1",
-                    "ImageId": EXAMPLE_AMI_ID,
-                },
-            },
-        },
-    }
-
-    client = boto3.client("ec2", region_name="us-east-1")
-    cf_conn = boto3.client("cloudformation", region_name="us-east-1")
-    cf_conn.create_stack(
-        StackName="test_stack", TemplateBody=json.dumps(dummy_template_json)
-    )
-    associations = client.describe_iam_instance_profile_associations()
-    associations["IamInstanceProfileAssociations"].should.have.length_of(1)
-    associations["IamInstanceProfileAssociations"][0]["IamInstanceProfile"][
-        "Arn"
-    ].should.contain("test_stack")
-
-    cf_conn.delete_stack(StackName="test_stack")
-    associations = client.describe_iam_instance_profile_associations()
-    associations["IamInstanceProfileAssociations"].should.have.length_of(0)
